Pass int to core interface

+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/LoginInteractor.java
@@ -17,10 +17,10 @@ public class LoginInteractor extends BaseLoginInteractor implements BaseLoginCon
     @Override
     protected void scheduleJobsPeriodically() {
         LocationTaskServiceJob.scheduleJob(LocationTaskServiceJob.TAG,
-                BuildConfig.SYNC_INTERVAL_IN_MINUTES, getFlexValue(BuildConfig.SYNC_INTERVAL_IN_MINUTES));
+                BuildConfig.SYNC_INTERVAL_IN_MINUTES, getFlexValue((int) BuildConfig.SYNC_INTERVAL_IN_MINUTES));
 
         PullUniqueIdsServiceJob.scheduleJob(SyncServiceJob.TAG,
-                BuildConfig.PULL_UNIQUE_IDS_MINUTES, getFlexValue(BuildConfig.PULL_UNIQUE_IDS_MINUTES));
+                BuildConfig.PULL_UNIQUE_IDS_MINUTES, getFlexValue((int) BuildConfig.PULL_UNIQUE_IDS_MINUTES));
     }
 
     @Override
